macho: add first pass at allocating parsed atoms in objects
This commit makes it possible to combine self-hosted with a pre-compiled
C object file, e.g.:

```
zig-out/bin/zig build-exe hello.zig add.o
```

where `add.o` is a pre-compiled C object file.

+++ src/link/MachO/commands.zig
@@ -337,7 +337,7 @@ pub const SegmentCommand = struct {
         return null;
     }
 
-    pub fn findFreeSpace(self: SegmentCommand, object_size: u64, min_alignment: u16, start: ?u64) u64 {
+    pub fn findFreeSpace(self: SegmentCommand, object_size: u64, min_alignment: u32, start: ?u64) u64 {
         var st: u64 = if (start) |v| v else self.inner.fileoff;
         while (self.detectAllocCollision(st, object_size)) |item_end| {
             st = mem.alignForwardGeneric(u64, item_end, min_alignment);
